Added data transformer and internal data container

+++ gradient_boosting/categories/CategoricalConverter.h
@@ -12,14 +12,14 @@ namespace categories {
 class CategoricalConverter {
  public:
   explicit CategoricalConverter(const std::vector<std::string>& features,
-                                const std::vector<size_t>& class_ids);
+                                const std::vector<double>& class_ids);
   std::vector<double> GetConversionResult() const;
   double Convert(const std::string& feature_value) const;
  private:
   double Convert(size_t id) const;
 
   double default_probability_;
-  std::vector<size_t> class_sum_;
+  std::vector<double> class_sum_;
   const CategoricalContainer container_;
   std::vector<double> conversion_result_;
   std::vector<size_t> features_count_;
